requirements.txt
notion-client
discord.py
python-dotenv
boto3
httpcore[asyncio]